+++ HISTORY.rst
@@ -47,4 +47,6 @@ History
 ------------------
 * added a function to revert applied filters
 * added a function that prints the current pipeline
+* added a function that clears all transformations
 * fixed a bug in which ``tabulate_flat_list_of_dicts`` would fail if the dictionary was empty
+* ``self.aggregation`` does not hold a copy of the files anymore but references to ``self.files``